Added a class for Project Modal Window
